refactor: rename search
