+++ src/assets/scss/explore.scss
