+++ src/components/SearchBar.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import './../assets/scss/searchbar.scss'
+
+const SearchBar = () => {
+    return (
+        <div className='search-bar'>
+            Search Bar
+        </div>
+    )
+}
+export default SearchBar
