--- src/components/SerchBar.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import './../asset/scss/serchbar.scss'
-
-const SerchBar = () => {
-    return (
-        <div className='serch-bar'>
-            Serch Bar
-        </div>
-    )
-}
-export default SerchBar
