+++ src/index.js
@@ -7,7 +7,7 @@ import { createStore, applyMiddleware } from 'redux'
 import { Provider } from 'react-redux'
 import logger from 'redux-logger'
 import reducer from './reducers'
-import './asset/css/index.css'
+import './assets/css/index.css'
 import Routes from './router'
 
 // create redux store
