+++ src/pages/Home.js
@@ -1,14 +1,14 @@
 import React from 'react'
-import '../asset/scss/home.scss'
+import '../assets/scss/home.scss'
 import NavBar from '../components/Navbar'
-import SerchBar from '../components/SerchBar'
+import SearchBar from '../components/SearchBar'
 import Explore from '../components/Explore'
 import NearBy from '../components/NearBy'
 
 const Home = () => (
     <div className='home-page'>
         <NavBar />
-        <SerchBar />
+        <SearchBar />
         <Explore />
         <NearBy />
     </div>
